Add GetMany member references for IListAdapter and IReadOnlyListAdapter
Introduces methods to retrieve MemberReference for GetMany on IListAdapter<T> and IReadOnlyListAdapter<T> in InteropReferences. Updates IList1 and IReadOnlyList1 builders to use the new GetMany references instead of GetAt, ensuring correct method binding for 'GetMany' implementations.

+++ src/WinRT.Runtime2/InteropServices/Collections/IListAdapter{T}.cs
@@ -225,7 +225,7 @@ public static class IListAdapter<T>
     /// <param name="items">The target <see cref="Span{T}"/> to write items into.</param>
     /// <returns>The number of items that were retrieved. This value can be less than the size of <paramref name="items"/> if the end of the list is reached.</returns>
     /// <see href="https://learn.microsoft.com/uwp/api/windows.foundation.collections.ivector-1.getmany"/>
-    public static int GetMany(IList<T> list, uint startIndex, Span<T> items)
+    public static uint GetMany(IList<T> list, uint startIndex, Span<T> items)
     {
         int count = list.Count;
 
@@ -251,6 +251,6 @@ public static class IListAdapter<T>
             items[i] = list[i + (int)startIndex];
         }
 
-        return itemCount;
+        return (uint)itemCount;
     }
 }
